+++ RunTime/Toolkits/Core/ActionKit/Framework/ActionQueueRecycleCallback.cs
@@ -13,7 +13,7 @@ namespace Framework3.Toolkits.ActionKit
     public struct ActionQueueRecycleCallback<T> : IActionQueueCallback where T : class, IAction
     {
         /// <summary>
-        /// 需要回收倒哪个 Pool 中
+        /// 需要回收倒哪个 ObjectPool 中
         /// </summary>
         private ObjectPool<T> _pool;