+++ Samples/Toolkits/AudioKit/0.Basic/AudioExample.cs
@@ -8,16 +8,12 @@
 
 namespace Framework3.Toolkits.AudioKit.Example._0.Basic
 {
+    using PoolKit;
     using UnityEngine;
     using UnityEngine.UI;
 
     public class AudioExample : MonoBehaviour
     {
-        private Button _btnPlaySound;
-        private Button _btnPlayMusic;
-        private Button _btnPlayNarration;
-        private Button _btnStopAllSound;
-
         private Slider _sldMusicVolume;
         private Slider _sldNarrationVolume;
         private Slider _sldSoundVolume;
@@ -28,12 +24,8 @@ namespace Framework3.Toolkits.AudioKit.Example._0.Basic
 
         private void Awake()
         {
-            // AudioKit.SoundMode = SoundMode.IgnoreSameSoundInSoundFrames;
-            
-            _btnPlaySound     = GameObject.Find("Canvas/BtnPlaySound").GetComponent<Button>();
-            _btnPlayMusic     = GameObject.Find("Canvas/BtnPlayMusic").GetComponent<Button>();
-            _btnPlayNarration = GameObject.Find("Canvas/BtnPlayNarration").GetComponent<Button>();
-            _btnStopAllSound  = GameObject.Find("Canvas/BtnStopAllSound").GetComponent<Button>();
+            // 使用 ResourcesAudioLoader 作为音频加载器
+            AudioKit.AudioLoaderPool.SetObjectFactory(new CustomObjectFactory<IAudioLoader>(() => new ResourcesAudioLoader()));
 
             _sldMusicVolume     = GameObject.Find("Canvas/Music/Slider").GetComponent<Slider>();
             _sldNarrationVolume = GameObject.Find("Canvas/Narration/Slider").GetComponent<Slider>();
@@ -42,65 +34,6 @@ namespace Framework3.Toolkits.AudioKit.Example._0.Basic
             _togMusicOn     = GameObject.Find("Canvas/Music/Toggle").GetComponent<Toggle>();
             _togNarrationOn = GameObject.Find("Canvas/Narration/Toggle").GetComponent<Toggle>();
             _togSoundOn     = GameObject.Find("Canvas/Sound/Toggle").GetComponent<Toggle>();
-
-            _btnPlaySound.onClick.AddListener(() =>
-            {
-                AudioKit.PlaySound(
-                    "ButtonClicked",
-                    onPlayFinish: player =>
-                    {
-                        Debug.Log("Sound play finish: " + player.AudioClipName);
-                    }
-                );
-            });
-            _btnPlayMusic.onClick.AddListener(() =>
-            {
-                AudioKit.PlayMusic(
-                    "PillowTalk",
-                    onPlayFinish: repeatCount =>
-                    {
-                        Debug.Log("Music play finish");
-                    });
-            });
-            _btnPlayNarration.onClick.AddListener(() =>
-            {
-                AudioKit.PlayNarration(
-                    "HomeBg",
-                    onPlayFinish: repeatCount =>
-                    {
-                        Debug.Log("Narration play finish");
-                    });
-            });
-            _btnStopAllSound.onClick.AddListener(() =>
-            {
-                AudioKit.StopAllSound();
-            });
-
-            _sldMusicVolume.onValueChanged.AddListener((value) =>
-            {
-                AudioKit.Setting.MusicVolume.Value = value;
-            });
-            _sldNarrationVolume.onValueChanged.AddListener((value) =>
-            {
-                AudioKit.Setting.NarrationVolume.Value = value;
-            });
-            _sldSoundVolume.onValueChanged.AddListener((value) =>
-            {
-                AudioKit.Setting.SoundVolume.Value = value;
-            });
-
-            _togMusicOn.onValueChanged.AddListener((value) =>
-            {
-                AudioKit.Setting.IsMusicOn.Value = value;
-            });
-            _togNarrationOn.onValueChanged.AddListener((value) =>
-            {
-                AudioKit.Setting.IsNarrationOn.Value = value;
-            });
-            _togSoundOn.onValueChanged.AddListener((value) =>
-            {
-                AudioKit.Setting.IsSoundOn.Value = value;
-            });
         }
 
         private void Start()
@@ -112,6 +45,81 @@ namespace Framework3.Toolkits.AudioKit.Example._0.Basic
             _togMusicOn.isOn     = AudioKit.Setting.IsMusicOn;
             _togNarrationOn.isOn = AudioKit.Setting.IsNarrationOn;
             _togSoundOn.isOn     = AudioKit.Setting.IsSoundOn;
+        }
+
+        public void PlayMusic()
+        {
+            Debug.Log("Playing Music...");
+            AudioKit.PlayMusic(
+                "Music/PillowTalk",
+                onPlayFinish: repeatCount =>
+                {
+                    Debug.Log("Music play finished.");
+                });
+        }
+        
+        public void PlayNarration()
+        {
+            Debug.Log("Playing Narration...");
+            AudioKit.PlayNarration(
+                "Music/HomeBg",
+                onPlayFinish: repeatCount =>
+                {
+                    Debug.Log("Narration play finished.");
+                });
+        }
+        
+        public void PlaySound()
+        {
+            Debug.Log("Playing Sound...");
+            AudioKit.PlaySound(
+                "Sound/ButtonClicked",
+                onPlayFinish: player =>
+                {
+                    Debug.Log("Sound play finished: " + player.AudioClipName);
+                });
+        }
+        
+        public void StopAllSound()
+        {
+            Debug.Log("Stopping All Sound...");
+            AudioKit.StopAllSound();
+        }
+        
+        public void SetMusicVolume(float volume)
+        {
+            AudioKit.Setting.MusicVolume.Value = volume;
+            Debug.Log("Music Volume set to: " + volume);
+        }
+        
+        public void SetNarrationVolume(float volume)
+        {
+            AudioKit.Setting.NarrationVolume.Value = volume;
+            Debug.Log("Narration Volume set to: " + volume);
+        }
+        
+        public void SetSoundVolume(float volume)
+        {
+            AudioKit.Setting.SoundVolume.Value = volume;
+            Debug.Log("Sound Volume set to: " + volume);
+        }
+        
+        public void ToggleMusic(bool isOn)
+        {
+            AudioKit.Setting.IsMusicOn.Value = isOn;
+            Debug.Log("Music toggled: " + isOn);
+        }
+        
+        public void ToggleNarration(bool isOn)
+        {
+            AudioKit.Setting.IsNarrationOn.Value = isOn;
+            Debug.Log("Narration toggled: " + isOn);
+        }
+        
+        public void ToggleSound(bool isOn)
+        {
+            AudioKit.Setting.IsSoundOn.Value = isOn;
+            Debug.Log("Sound toggled: " + isOn);
         }
     }
 }