+++ Samples/Toolkits/Core/PoolKit/0.ObjectPoolExample/ObjectPoolExample.cs
@@ -8,46 +8,63 @@
 
 namespace Framework3.Toolkits.PoolKit.Example._0.ObjectPoolExample
 {
-    using System.Collections;
-    using System.Collections.Generic;
-    using Sirenix.OdinInspector;
     using UnityEngine;
+    using UnityEngine.UI;
 
     public class ObjectPoolExample : MonoBehaviour
     {
-        [ShowInInspector]
         private ObjectPool<GameObject> _objectPool;
+
+        [SerializeField]
+        private Image _imageItem;
+
+        [SerializeField]
+        private Transform _poolParent;
+
+        [SerializeField]
+        private Transform _useParent;
 
         void Start()
         {
             _objectPool = new ObjectPool<GameObject>(
                 () =>
                 {
-                    var gameObj = new GameObject();
-                    gameObj.SetActive(false);
-                    return gameObj;
+                    var image = Instantiate(_imageItem, _poolParent);
+                    image.gameObject.SetActive(true);
+                    image.color = Color.gray;
+                    return image.gameObject;
                 },
-                actionOnRelease: gameObj => { gameObj.SetActive(false); },
+                actionOnGet: gameObj =>
+                {
+                    gameObj.transform.SetParent(_useParent);
+                    gameObj.GetComponent<Image>().color = Color.white;
+                },
+                actionOnRelease: gameObj =>
+                {
+                    gameObj.transform.SetParent(_poolParent);
+                    gameObj.GetComponent<Image>().color = Color.gray;
+                },
                 actionOnDestroy: Destroy,
                 defaultCapacity: 10);
         }
-        
+
         public void GetGameObjectFromPool()
         {
-            var obj = _objectPool.Get();
-            obj.SetActive(true);
-            StartCoroutine(Recycle(obj));
+            _objectPool.Get();
         }
-        
+
+        public void ReleaseGameObjectToPool()
+        {
+            if (_useParent.childCount > 0)
+            {
+                var item = _useParent.GetChild(0);
+                _objectPool.Release(item.gameObject);
+            }
+        }
+
         public void ClearGameObjectPool()
         {
             _objectPool.Clear();
-        }
-
-        private IEnumerator Recycle(GameObject obj)
-        {
-            yield return new WaitForSeconds(1);
-            _objectPool.Release(obj);
         }
     }
 }